requirements.txt
flask==3.1.0
werkzeug==3.1.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-distro==0.50b0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-instrumentation-flask==0.50b0
